# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CreateInboxShareRequest {
  /**
     * @minLength 1
     * @maxLength 80
     */
  senderName: string;
  /** @maxLength 50 */
  message: string;
  /**
     * @minLength 1
     * @maxLength 200
     */
  notificationTitle: string;
  /**
     * @minLength 1
     * @maxLength 2000
     */
  notificationBody: string;
  actionUrl?: string | null;
}

export interface InboxShare {
  token: string;
  url: string;
  createdAt: string;
}

export interface SharedNotification {
  token: string;
  senderName: string;
  message: string;
  notificationTitle: string;
  notificationBody: string;
  actionUrl?: string | null;
  createdAt: string;
  claimed: boolean;
}

export interface InboxItem {
  id: string;
  token: string;
  senderName: string;
  message: string;
  notificationTitle: string;
  notificationBody: string;
  actionUrl?: string | null;
  createdAt: string;
  readAt?: string | null;
}

export interface InboxList {
  items: InboxItem[];
}