# lib/api-zod/src/generated/types/createInboxShareRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateInboxShareRequest {
  /**
     * @minLength 1
     * @maxLength 80
     */
  senderName: string;
  /** @maxLength 50 */
  message: string;
  /**
     * @minLength 1
     * @maxLength 200
     */
  notificationTitle: string;
  /**
     * @minLength 1
     * @maxLength 2000
     */
  notificationBody: string;
  actionUrl?: string | null;
}