# lib/api-zod/src/generated/types/inboxItem.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface InboxItem {
  id: string;
  token: string;
  senderName: string;
  message: string;
  notificationTitle: string;
  notificationBody: string;
  actionUrl?: string | null;
  createdAt: Date;
  readAt?: Date | null;
}