# lib/api-zod/src/generated/types/inboxShare.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface InboxShare {
  token: string;
  url: string;
  createdAt: Date;
}