# lib/api-zod/src/generated/types/sharedNotification.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SharedNotification {
  token: string;
  senderName: string;
  message: string;
  notificationTitle: string;
  notificationBody: string;
  actionUrl?: string | null;
  createdAt: Date;
  claimed: boolean;
}